fixed non-standard html,added messages block
putting anything that is possibly going to display anything to the user between the closing `</head>` and opening `<body>` tags is not a very clean or standard way of doing things, its better to have the markup rendered inside of the body tag where it belongs, also so users can override the placement of the flashed messages in their templates, its helpful to add that chunk of code into its own jinja block. your welcome.